差綠界跟coupon api

+++ Cat_Paw_Footprint/Areas/Notification/Services/CouponExpiryChecker.cs
@@ -58,7 +58,7 @@ namespace Cat_Paw_Footprint.Services
                 if (r.CustomerID > 0)
                 {
                     await notifSvc.AddNotificationAsync(
-                        r.CustomerID,
+                        (int)r.CustomerID,
                         "優惠券即將到期",
                         $"您的優惠券「{r.Coupon.CouponDesc}」將於 {r.Coupon.EndDate:MM/dd} 到期，請盡快使用！",
                         "優惠券"
